Adam-v1/requirements.txt
torch
torchvision
numpy
glob
tqdm
sklearn
scikit-image
scikit-learn
simpleitk
scipy
timm
Pillow
pretrainedmodels
efficientnet_pytorch